Detection Ring
Detection ring has been implemented. The ring encircles the player and changes its' shape depending on the relative location of nearby detectable objects. The segment of the ring closest to the detected object will "bump" upwards to highlight the relative angle. The "bump" will also be grater the closer the object is to the ring's centre.