Tripwire
Tripwire has been implemented into the game.  When the tripwire detects an collider, by using a raycast, an event is called which can be used to invoke an array of actions. The proximity sensor has been modified to find the point between the start and end of the tripwire that is nearest to the sensor's transform component.